requirements/requirements.txt
Django==4.1.5
django_extensions==3.2.1
django-environ==0.9.0
psycopg2==2.9.5
django-leaflet==0.28.3

djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1

channels[daphne]==4.0.0
channels_redis==4.0.0

ipython==8.8.0
